pyproject.toml
[project]
name = "modular-agent"
version = "0.1.0"
description = "Modular Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.3.1",
    "chromadb>=1.3.7",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-tavily>=0.2.14",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.1",
]
